+++ Sources/InstanaAgent/Beacons/Beacons Types/Beacon.swift
@@ -6,7 +6,7 @@ import Foundation
 
 /// Base class for Beacon.
 class Beacon: Identifiable {
-    let id = UUID()
+    let id = Beacon.generateUniqueIdImpl()
     var timestamp: Instana.Types.Milliseconds = Date().millisecondsSince1970
     let viewName: String?
 
@@ -14,6 +14,17 @@ class Beacon: Identifiable {
          viewName: String? = nil) {
         self.timestamp = timestamp
         self.viewName = InstanaProperties.validate(view: viewName)
+    }
+
+    // A hex encoded 64 bit random ID.
+    static func generateUniqueIdImpl() -> String {
+        let validIdCharacters = Array("0123456789abcdef")
+        var result: String = ""
+        for _ in 0 ..< 16 {
+            let idx = Int.random(in: 0 ... 15)
+            result += String(validIdCharacters[idx])
+        }
+        return result
     }
 }